Added Generate Code to Admin dashboard, made generate Code into a Component
