+++ src/routes/dashboard/administrator/+page.svelte
@@ -2,6 +2,7 @@
     import ProfilePicture from "$lib/components/ProfilePicture.svelte";
     import RegularButton from "$lib/components/RegularButton.svelte";
 	import DashboardNavigationCard from "$lib/components/DashboardNavigationCard.svelte";
+	import GenerateCodeCard from "$lib/components/GenerateCodeCard.svelte";
 
 	const links = [
 		{ name: "Manage Classes", href: "/dashboard" },
@@ -16,12 +17,7 @@
 
 			<h2 class="text-xl font-semibold">Marilin Monroe</h2>
 		</div>
-
-		<div class="w-1/2 flex-1 flex flex-col justify-center items-start pl-16 gap-4">
-			{#each links as link}
-				<RegularButton href={link.href} name={link.name}></RegularButton>
-			{/each}
-		</div>
+		<GenerateCodeCard></GenerateCodeCard>
 	</div>
 	<div class="w-full grid grid-cols-2 gird-columns-2 p-6 gap-6">
 		<DashboardNavigationCard 
@@ -48,5 +44,29 @@
 		buttonName="to Student feedback" 
 		href="fcgv"
 		/>
+		<DashboardNavigationCard 
+		description="View the student's feedback on assets and designs" 
+		title="Manage Assets" 
+		buttonName="to Student feedback" 
+		href="fcgv"
+		/>
+		<DashboardNavigationCard 
+		description="View the student's feedback on assets and designs" 
+		title="Manage Layouts" 
+		buttonName="to Student feedback" 
+		href="fcgv"
+		/>
+		<DashboardNavigationCard 
+		description="View the student's feedback on assets and designs" 
+		title="Manage Classes" 
+		buttonName="to Student feedback" 
+		href="fcgv"
+		/>
 	</div>
+
+	<!-- <div class="w-1/2 flex-1 flex flex-col justify-center items-start pl-16 gap-4">
+			{#each links as link}
+				<RegularButton href={link.href} name={link.name}></RegularButton>
+			{/each}
+		</div> -->
 </div>
